backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
Pillow==10.1.0
python-decouple==3.8
django-filter==23.5
django-ratelimit==4.1.0
drf-spectacular==0.27.0
psutil==5.9.6
gunicorn==21.2.0
